Flytter alle komponenter som blir brukt av LagBehandlingModal inn i tilhørende mappe. Rename: LagBehandlingModal -> OpprettBehandlingModal
